hostel_admin_backend/requirements.txt
django>=5.0
djangorestframework
django-filter
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
PyJWT
requests
python-decouple
setuptools
dj-database-url
Pillow
